gve/boot: fix bare-metal clippy (unnecessary_cast, too_many_args, doc)
These slipped past the host build / pre-push because os:none crates are
incompatible on the host platform and never get clippy-checked there;
they only surfaced when rust-analyzer's discover script built under the
unikernel target.

- gve/dqo.rs: drop a no-op u32->u32 cast; allow too_many_arguments on
  emit_tso_descs (8 args, matching the quic/tx.rs precedent).
- boot/entry.rs: DEEP_IDLE_ROUNDS had been inserted between idle_cb's
  doc comment and the function, merging two doc blocks (x86-only
  doc_lazy_continuation) and leaving idle_cb undocumented. Reorder so
  each doc attaches to its own item.

+++ crates/drivers/gve/src/dqo.rs
@@ -625,7 +625,7 @@ fn emit_ctx_pkt(
 #[inline]
 fn pkt_bounce_buf(tx: &TxQueue, fill_cnt: u32) -> (u64, u64) {
     let mask = (tx.ring_entries - 1) as u32;
-    let slot = (fill_cnt.wrapping_add(1) & mask) as u32;
+    let slot = fill_cnt.wrapping_add(1) & mask;
     let buf_offset = slot * (RX_BUFFER_SIZE as u32);
     (
         tx.qpl_base_va + buf_offset as u64,
@@ -888,6 +888,7 @@ pub(crate) fn submit_tx_tso(
 /// Emit the TSO descriptor burst (TSO-ctx, general-ctx, SG pkt descs)
 /// for a super-segment already in big-pool slot `big_slot` at DMA
 /// address `buf_phys`, advance `fill_cnt`, doorbell unless deferred.
+#[allow(clippy::too_many_arguments)]
 fn emit_tso_descs(
     tx: &TxQueue,
     qp: usize,
